% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{areaMesh}
\alias{areaMesh}
\title{Find the area of the Patch and create a mesh grid}
\usage{
areaMesh(xycord, ds = 0.1)
}
\arguments{
\item{data}{The data set with photo counts}

\item{type}{Type of correction. Use 1 for separate correction for each reader (assumes two readers) and 2 for the same correction for both readers}

\item{population}{Specify which population or populations to estimate the uncertainty from}

\item{readers}{Specify the ID of each reader. Assumes 2 readers}

\item{transect}{Vector containing the transect each photo comes from}

\item{Spacing}{Transect width}

\item{gap}{Distance threshold for length if flown over open water}
}
\value{
A mesh grid for used to estimate the GAM density on
}
\description{
Find the area of the Patch and create a mesh grid
}
\examples{
areaMesh()
}
