% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{birthDist}
\alias{birthDist}
\title{Estimate the temporal distribution of births}
\usage{
birthDist(
  harpfname = "HarpStages2012.txt",
  harpLengthStages = c(2.4, 4.42, 11.39),
  harpKappa = 12.4,
  hoodedfname = "HoodStages2012.txt",
  hoodedLengthStages = c(2, 1, 4),
  hoodedKappa = 8.6,
  survey = "WestIce2012",
  datePhoto = 28,
  Nsim = 10000,
  plotCI = FALSE,
  grDev = TRUE
)
}
\arguments{
\item{harpfname}{Filename of the harp seal staging data}

\item{harpLengthStages}{Length of each predefined harp seal stage}

\item{harpKappa}{A shape parameter defining the duration of each stage}

\item{hoodedfname}{Filename of the hooded seal staging data}

\item{hoodedKappa}{A shape parameter defining the duration of each stage}

\item{survey}{Name of the survey to be analyzed}

\item{datePhoto}{Date for which the photographic survey was carrie out}

\item{Nsim}{Number of Monte Carlo simulations used to calculate the 95% CI}

\item{plotCI}{Plot with or without 95% CI. Default is FALSE}

\item{grDev}{Load a graphical device}

\item{hoodedStages}{Length of each predefined hooded seal stage}
}
\value{
returnList A list containing the estimated mean and std of the birth distribution, the estimated proportions of seals on ice at the time of the photographic survey, and the estimated model fits to the observed staging data.
}
\description{
Estimating the temporal distribution of births to correct for pups not born yet or have left the ice. It is important to know that both populations cannot be run at the same time. Each population has to be run separately.
}
\examples{
birtDist()
}
