% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_oo.R
\name{dbinom2}
\alias{dbinom2}
\title{Binomial PDF}
\usage{
dbinom2(x, size, prob, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{size}{number of trials (zero or more). Unlike base 'dbinom' doesn't have to be an integer}

\item{prob}{probability of success on each trial}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
A binomial pdf that handles a noninteger 'size' argument.
}
\author{
Devin S. Johnson
}
