% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_oo.R
\name{fit_oo}
\alias{fit_oo}
\title{Fit observer by occasion model to NFS resight data}
\usage{
fit_oo(data, M, start_val = NULL, refit = TRUE, ...)
}
\arguments{
\item{data}{A data frame containing the resight information. The data must contain the columns 
(1) 'obs'-- the categorical name of the observer.
(2) 'resample'-- a variable indicating the resample occasion
(3) 'm'-- the counts of marked pups by each observer
(4) 'u'-- the counts of unmarked pups}

\item{M}{the total number of unmarked pups in the population}

\item{start_val}{an optional vector of starting values for the model}

\item{refit}{logical. should the function attempt a refit to help  ensure global optimum is reached}

\item{...}{holding space for any other variable supplied, these will be ignored}
}
\description{
Fit the observer by occasion model using maximum likelihood. The parameters, covariance matrix, 
and log-likelihood are returned in a list.
}
