% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{GAMestimate}
\alias{GAMestimate}
\title{Estimate the pup production using GAMs}
\usage{
GAMestimate(
  harpcounts = NA,
  hoodedcounts = NA,
  area,
  xycord,
  transect,
  population = c("harp", "hood"),
  distr = "negbin",
  Spacing = 3,
  ds = 0.1,
  plotMesh = TRUE,
  grDev = TRUE
)
}
\arguments{
\item{harpcounts}{The data set with photo counts of harp seals}

\item{hoodedcounts}{The data set with photo counts of harp seals}

\item{area}{Define the area used for estimating the abundance}

\item{transect}{Vector containing the transect each photo comes from}

\item{population}{Specify which population or populations to estimate the abundance}

\item{distr}{Distance threshold for length if flown over open water}

\item{ds}{Resolution of the grid to estimate the GAM surface}

\item{grDev}{Logical parameter to turn on and off wether to plot the results in a graphical device}
}
\value{
A list containing the estimates obtained by the GAM method
}
\description{
Estimate the pup production using GAMs
}
\examples{
GAMestimate()
}
