% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_oo.R
\name{get_IS_sample}
\alias{get_IS_sample}
\title{Draw posterior sample using Sample-Importance-Resample algorithm}
\usage{
get_IS_sample(fit, data, M, res = 5000, initial = 10000, ...)
}
\arguments{
\item{fit}{A list produced by `fit_oo' function}

\item{data}{The resampling data}

\item{M}{the total number of pups shear marked}

\item{res}{the size of the resample}

\item{initial}{size of the initial sample from the MVN proposal}

\item{...}{holding space for any other variable supplied, these will be ignored}
}
\description{
Uses SIR methodology to draw a sample from the posterior distribution. A MVN proposal is used based on the MLE list output of the model fitting function
}
