% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{kingsley}
\alias{kingsley}
\title{Estimates the pup production using the Kingsley method}
\usage{
kingsley(x, count, area, transect, Spacing = 3, gap = 1)
}
\arguments{
\item{x}{X Coordinates in Cartesian system}

\item{count}{Vector containing the pup counts for each photo}

\item{area}{Vector containing area of each photo}

\item{transect}{Vector containing the transect each photo comes from}

\item{Spacing}{Transect width in NM, default 3}

\item{gap}{Maximum distance in NM between allowed, for removing holes where camera has been stopped due to flying over open water, default = 1}
}
\value{

}
\description{
Estimates the pup production using the Kingsley method
}
\examples{
kingsley()
}
