% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lb2xy}
\alias{lb2xy}
\title{Convert latitude and longitude to Cartesian koordinates}
\usage{
lb2xy(lon, lat, lon0, lat0)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}

\item{lon0}{Mean longitude}

\item{lat0}{Mean latitude}
}
\value{

}
\description{
Convert latitude and longitude to Cartesian koordinates
}
\examples{
lb2xy()
}
