% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_oo.R
\name{ln_lik_oo}
\alias{ln_lik_oo}
\title{log likelihood funtion for observer by occasion model}
\usage{
ln_lik_oo(par, M, u, m)
}
\arguments{
\item{par}{parameter vector}

\item{M}{Number of marks distributed}

\item{u}{vector on unmarked animals counted by each observer on each occasion}

\item{m}{vector on marked animals counted by each observer on each occasion}
}
\description{
calculate log-likelihood for the observer by occassion model
}
\author{
Devin S. Johnson
}
