% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plotGAM}
\alias{plotGAM}
\title{Plot GAM fit do data}
\usage{
plotGAM(seal.mod, x, y, transect, area, spacing, patch)
}
\arguments{
\item{x}{Vector containing x coordinates in Cartesian system}

\item{y}{Vector containing y coordinates in Cartesian system}

\item{transect}{Transect each photo beongs to}

\item{area}{Area of each photo}

\item{spacing}{Distance between transects}

\item{patch}{?}
}
\value{

}
\description{
Plot GAM fit do data
}
\examples{
plotGAM()
}
