% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{printKingsley2screen}
\alias{printKingsley2screen}
\title{Print pup production estimates using the Kingsley method to screen}
\usage{
printKingsley2screen(estimates, population = c("harp", "hood"))
}
\arguments{
\item{population}{Specify which population or populations to estimate the uncertainty from}

\item{data}{The data set with photo counts}

\item{type}{Type of correction. Use 1 for separate correction for each reader (assumes two readers) and 2 for the same correction for both readers}

\item{readers}{Specify the ID of each reader. Assumes 2 readers}

\item{transect}{Vector containing the transect each photo comes from}

\item{Spacing}{Transect width}

\item{gap}{Distance threshold for length if flown over open water}
}
\value{

}
\description{
Print pup production estimates using the Kingsley method to screen
}
\examples{
print2screen()
}
