% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{propFitFun}
\alias{propFitFun}
\title{Calculates the fitted staging curves}
\usage{
propFitFun(mub = NA, sigmab = NA, data = NA, timeDF = NA)
}
\arguments{
\item{mub}{Mean of the birth distribution}

\item{sigmab}{Sigma for det birth distribution}

\item{data}{The data set with photo counts}

\item{type}{Type of correction. Use 1 for separate correction for each reader (assumes two readers) and 2 for the same correction for both readers}
}
\value{
returnList A list containing the fittet proportion curves given the parameters of the birth distribution
}
\description{
Calcculates the fitted staging curves for a given birth distribution
}
\examples{
propFitFun()
}
