% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rottweileR.R
\name{rottweileR}
\alias{rottweileR}
\title{rottweileR}
\usage{
rottweileR()
}
\value{
A photo of a Rottweiler is downloaded and displayed in the plot window.
}
\description{
This function was inspired by and named for my Rottweiler, Einstein. These aren't pictures of Einstein. If you want to see those, check out einsteinstheoryofbarktivity on Instagram!
}
\examples{
rottweileR()

}
