% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{see_demo}
\alias{see_demo}
\title{View demonstration}
\usage{
see_demo()
}
\description{
This function allows users to access a full step-by-step guide for using 'pupR' and other packages in the
 R statistical environment for estimating northern fur seal pup production.
}
