% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_oo.R
\name{summarize_oo}
\alias{summarize_oo}
\title{Summarize the posterior sample}
\usage{
summarize_oo(SIR, M, ...)
}
\arguments{
\item{SIR}{a matrix of posterior samples from the `get_IS_sample' function}

\item{M}{total number of sheared pups}

\item{...}{holding space for any other variable supplied, these will be ignored}
}
\description{
Here we summarize the posterior sample into a table for each rookery
}
