% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{vmeasest}
\alias{vmeasest}
\title{Estimate uncertainty from measurement errors}
\usage{
vmeasest(
  x,
  count,
  area,
  transect,
  Spacing,
  gap,
  var_b,
  var_u,
  var_a,
  cov_ab = NA,
  intcpt = FALSE
)
}
\arguments{
\item{x}{X Coordinates in Cartesian system}

\item{count}{Vector containing the pup counts for each photo}

\item{area}{Vector containing area of each photo}

\item{transect}{Vector containing the transect each photo comes from}

\item{Spacing}{Distance between transect}

\item{gap}{Distance threshold for length if flown over open water}
}
\value{

}
\description{
Estimate uncertainty from measurement errors
}
\examples{
vmeasest()
}
