% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{loadAndPrepareData}
\alias{loadAndPrepareData}
\title{Convert latitude and longitude to Cartesian koordinates}
\usage{
loadAndPrepareData(
  survey = "WestIce2012",
  fname = "WestIce2012.csv",
  population = c("harp", "hood"),
  camParamLength = 0.06786,
  CamParamWidth = 0.10386,
  CamParam = 0.1005
)
}
\arguments{
\item{survey}{Specify which survey to analyze}

\item{fname}{File name of data file}

\item{population}{Specify which population to be analysed, harp, hood or both.}
}
\value{

}
\description{
Convert latitude and longitude to Cartesian koordinates
}
\examples{
lb3xy()
}
