\name{pupR-package}
\alias{pupR-package}
\alias{pupR}
\docType{package}
\title{
\packageTitle{pupR}
}
\description{
\packageDescription{pupR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pupR}
\packageIndices{pupR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pupR}

Maintainer: \packageMaintainer{pupR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
